requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.39.0
datasets>=2.12.0

# CLIP and multimodal (via Hugging Face transformers)
# CLIP models are available through transformers library

# Graph and RAG
networkx>=3.0
pyvis>=0.3.2
langchain>=0.3.0
langchain-community>=0.3.0

# State Space Models (via Hugging Face)
# Note: State-space models are available through transformers library
# mamba-ssm>=1.0.0  # Optional: enables optimized CUDA kernels
# causal-conv1d>=1.2.0  # Optional: enables optimized CUDA kernels
# If not installed, the "eager" implementation will be used

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
scipy>=1.10.0
faiss-cpu>=1.7.0

# Community Detection (Microsoft GraphRAG-inspired)
python-louvain>=0.16
# leidenalg>=0.10.0  # Optional: Leiden algorithm (requires igraph)
# igraph>=0.10.0  # Optional: Advanced graph analysis

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
